requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-filter>=24.1
psycopg[binary]>=3.1
python-dotenv>=1.0
Pillow>=10.0
boto3>=1.28
django-storages[boto3]>=1.14
weasyprint>=61.0
qrcode>=7.4
reportlab>=4.0
pyphen>=0.14
arabic-reshaper>=3.0
python-bidi>=0.4
uvicorn>=0.24
django-cors-headers>=4.3
dj-database-url>=2.1
